# internal/web/tasks/api.gen.go
// Package tasks provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package tasks

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	strictecho "github.com/oapi-codegen/runtime/strictmiddleware/echo"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// RequestTaskBody defines model for RequestTaskBody.
type RequestTaskBody struct {
	Title  *string             `json:"title,omitempty"`
	UserId *openapi_types.UUID `json:"user_id,omitempty"`
}

// RequestTitleUpdate defines model for RequestTitleUpdate.
type RequestTitleUpdate struct {
	Title *string `json:"title,omitempty"`
}

// Task defines model for Task.
type Task struct {
	Id     *openapi_types.UUID `json:"id,omitempty"`
	IsDone *bool               `json:"is_done,omitempty"`
	Title  *string             `json:"title,omitempty"`
	UserId *openapi_types.UUID `json:"user_id,omitempty"`
}

// GetTasksParams defines parameters for GetTasks.
type GetTasksParams struct {
	IsDone *bool `form:"is_done,omitempty" json:"is_done,omitempty"`
}

// CreateTaskJSONRequestBody defines body for CreateTask for application/json ContentType.
type CreateTaskJSONRequestBody = RequestTaskBody

// UpdateTitleTaskByIdJSONRequestBody defines body for UpdateTitleTaskById for application/json ContentType.
type UpdateTitleTaskByIdJSONRequestBody = RequestTitleUpdate

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get All Tasks
	// (GET /tasks)
	GetTasks(ctx echo.Context, params GetTasksParams) error
	// Create Task
	// (POST /tasks)
	CreateTask(ctx echo.Context) error
	// Delete Task By ID
	// (DELETE /tasks/{id})
	DeleteTaskById(ctx echo.Context, id openapi_types.UUID) error
	// Get Task By ID
	// (GET /tasks/{id})
	GetTaskById(ctx echo.Context, id openapi_types.UUID) error
	// Update Task's Completed
	// (PATCH /tasks/{id})
	UpdateTaskCompletedById(ctx echo.Context, id openapi_types.UUID) error
	// Update Task's Title
	// (PUT /tasks/{id})
	UpdateTitleTaskById(ctx echo.Context, id openapi_types.UUID) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// GetTasks converts echo context to params.
func (w *ServerInterfaceWrapper) GetTasks(ctx echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetTasksParams
	// ------------- Optional query parameter "is_done" -------------

	err = runtime.BindQueryParameter("form", true, false, "is_done", ctx.QueryParams(), &params.IsDone)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter is_done: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetTasks(ctx, params)
	return err
}

// CreateTask converts echo context to params.
func (w *ServerInterfaceWrapper) CreateTask(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreateTask(ctx)
	return err
}

// DeleteTaskById converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteTaskById(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithLocation("simple", false, "id", runtime.ParamLocationPath, ctx.Param("id"), &id)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeleteTaskById(ctx, id)
	return err
}

// GetTaskById converts echo context to params.
func (w *ServerInterfaceWrapper) GetTaskById(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithLocation("simple", false, "id", runtime.ParamLocationPath, ctx.Param("id"), &id)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetTaskById(ctx, id)
	return err
}

// UpdateTaskCompletedById converts echo context to params.
func (w *ServerInterfaceWrapper) UpdateTaskCompletedById(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithLocation("simple", false, "id", runtime.ParamLocationPath, ctx.Param("id"), &id)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.UpdateTaskCompletedById(ctx, id)
	return err
}

// UpdateTitleTaskById converts echo context to params.
func (w *ServerInterfaceWrapper) UpdateTitleTaskById(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithLocation("simple", false, "id", runtime.ParamLocationPath, ctx.Param("id"), &id)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.UpdateTitleTaskById(ctx, id)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(baseURL+"/tasks", wrapper.GetTasks)
	router.POST(baseURL+"/tasks", wrapper.CreateTask)
	router.DELETE(baseURL+"/tasks/:id", wrapper.DeleteTaskById)
	router.GET(baseURL+"/tasks/:id", wrapper.GetTaskById)
	router.PATCH(baseURL+"/tasks/:id", wrapper.UpdateTaskCompletedById)
	router.PUT(baseURL+"/tasks/:id", wrapper.UpdateTitleTaskById)

}

type GetTasksRequestObject struct {
	Params GetTasksParams
}

type GetTasksResponseObject interface {
	VisitGetTasksResponse(w http.ResponseWriter) error
}

type GetTasks200JSONResponse []Task

func (response GetTasks200JSONResponse) VisitGetTasksResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type CreateTaskRequestObject struct {
	Body *CreateTaskJSONRequestBody
}

type CreateTaskResponseObject interface {
	VisitCreateTaskResponse(w http.ResponseWriter) error
}

type CreateTask201JSONResponse Task

func (response CreateTask201JSONResponse) VisitCreateTaskResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type DeleteTaskByIdRequestObject struct {
	Id openapi_types.UUID `json:"id"`
}

type DeleteTaskByIdResponseObject interface {
	VisitDeleteTaskByIdResponse(w http.ResponseWriter) error
}

type DeleteTaskById204Response struct {
}

func (response DeleteTaskById204Response) VisitDeleteTaskByIdResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type GetTaskByIdRequestObject struct {
	Id openapi_types.UUID `json:"id"`
}

type GetTaskByIdResponseObject interface {
	VisitGetTaskByIdResponse(w http.ResponseWriter) error
}

type GetTaskById200JSONResponse Task

func (response GetTaskById200JSONResponse) VisitGetTaskByIdResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type UpdateTaskCompletedByIdRequestObject struct {
	Id openapi_types.UUID `json:"id"`
}

type UpdateTaskCompletedByIdResponseObject interface {
	VisitUpdateTaskCompletedByIdResponse(w http.ResponseWriter) error
}

type UpdateTaskCompletedById200JSONResponse Task

func (response UpdateTaskCompletedById200JSONResponse) VisitUpdateTaskCompletedByIdResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type UpdateTitleTaskByIdRequestObject struct {
	Id   openapi_types.UUID `json:"id"`
	Body *UpdateTitleTaskByIdJSONRequestBody
}

type UpdateTitleTaskByIdResponseObject interface {
	VisitUpdateTitleTaskByIdResponse(w http.ResponseWriter) error
}

type UpdateTitleTaskById200JSONResponse Task

func (response UpdateTitleTaskById200JSONResponse) VisitUpdateTitleTaskByIdResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Get All Tasks
	// (GET /tasks)
	GetTasks(ctx context.Context, request GetTasksRequestObject) (GetTasksResponseObject, error)
	// Create Task
	// (POST /tasks)
	CreateTask(ctx context.Context, request CreateTaskRequestObject) (CreateTaskResponseObject, error)
	// Delete Task By ID
	// (DELETE /tasks/{id})
	DeleteTaskById(ctx context.Context, request DeleteTaskByIdRequestObject) (DeleteTaskByIdResponseObject, error)
	// Get Task By ID
	// (GET /tasks/{id})
	GetTaskById(ctx context.Context, request GetTaskByIdRequestObject) (GetTaskByIdResponseObject, error)
	// Update Task's Completed
	// (PATCH /tasks/{id})
	UpdateTaskCompletedById(ctx context.Context, request UpdateTaskCompletedByIdRequestObject) (UpdateTaskCompletedByIdResponseObject, error)
	// Update Task's Title
	// (PUT /tasks/{id})
	UpdateTitleTaskById(ctx context.Context, request UpdateTitleTaskByIdRequestObject) (UpdateTitleTaskByIdResponseObject, error)
}

type StrictHandlerFunc = strictecho.StrictEchoHandlerFunc
type StrictMiddlewareFunc = strictecho.StrictEchoMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// GetTasks operation middleware
func (sh *strictHandler) GetTasks(ctx echo.Context, params GetTasksParams) error {
	var request GetTasksRequestObject

	request.Params = params

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetTasks(ctx.Request().Context(), request.(GetTasksRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetTasks")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetTasksResponseObject); ok {
		return validResponse.VisitGetTasksResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// CreateTask operation middleware
func (sh *strictHandler) CreateTask(ctx echo.Context) error {
	var request CreateTaskRequestObject

	var body CreateTaskJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.CreateTask(ctx.Request().Context(), request.(CreateTaskRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "CreateTask")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(CreateTaskResponseObject); ok {
		return validResponse.VisitCreateTaskResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// DeleteTaskById operation middleware
func (sh *strictHandler) DeleteTaskById(ctx echo.Context, id openapi_types.UUID) error {
	var request DeleteTaskByIdRequestObject

	request.Id = id

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.DeleteTaskById(ctx.Request().Context(), request.(DeleteTaskByIdRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "DeleteTaskById")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(DeleteTaskByIdResponseObject); ok {
		return validResponse.VisitDeleteTaskByIdResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// GetTaskById operation middleware
func (sh *strictHandler) GetTaskById(ctx echo.Context, id openapi_types.UUID) error {
	var request GetTaskByIdRequestObject

	request.Id = id

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetTaskById(ctx.Request().Context(), request.(GetTaskByIdRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetTaskById")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetTaskByIdResponseObject); ok {
		return validResponse.VisitGetTaskByIdResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// UpdateTaskCompletedById operation middleware
func (sh *strictHandler) UpdateTaskCompletedById(ctx echo.Context, id openapi_types.UUID) error {
	var request UpdateTaskCompletedByIdRequestObject

	request.Id = id

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.UpdateTaskCompletedById(ctx.Request().Context(), request.(UpdateTaskCompletedByIdRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "UpdateTaskCompletedById")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(UpdateTaskCompletedByIdResponseObject); ok {
		return validResponse.VisitUpdateTaskCompletedByIdResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// UpdateTitleTaskById operation middleware
func (sh *strictHandler) UpdateTitleTaskById(ctx echo.Context, id openapi_types.UUID) error {
	var request UpdateTitleTaskByIdRequestObject

	request.Id = id

	var body UpdateTitleTaskByIdJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.UpdateTitleTaskById(ctx.Request().Context(), request.(UpdateTitleTaskByIdRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "UpdateTitleTaskById")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(UpdateTitleTaskByIdResponseObject); ok {
		return validResponse.VisitUpdateTitleTaskByIdResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}