# openapi/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: API
paths:
  /tasks:
    get:
      summary: Get All Tasks
      operationId: GetTasks
      tags:
        - tasks
      parameters:
        - in: query
          name: is_done
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Finded Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        
    post:
      summary: Create Task
      operationId: CreateTask
      tags:
        - tasks
      requestBody:
        description: Task Data For Create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestTaskBody'
      responses:
        '201':
          description: The Task Was Successfully Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /tasks/{id}:
    get:
      summary: Get Task By ID
      operationId: GetTaskById
      tags:
        - tasks
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User By ID Was Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
      
    patch: 
      summary: Update Task's Completed
      operationId: UpdateTaskCompletedById
      tags:
        - tasks
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Updated Completed Task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
      
    put:
      summary: Update Task's Title
      operationId: UpdateTitleTaskById
      tags:
        - tasks
      parameters:
        - in: path
          name: id
          required: true
          schema: 
            type: string
            format: uuid
      requestBody:
        description: Title to Update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestTitleUpdate'
      responses:
        '200':
          description: Task Was Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
      
    delete:
      summary: Delete Task By ID
      operationId: DeleteTaskById
      tags:
        - tasks
      parameters: 
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Task Was Successfully Deleted
...
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        is_done:
          type: boolean
        user_id:
          type: string
          format: uuid
    RequestTaskBody:
      type: object
      properties:
        title:
          type: string
        user_id:
          type: string
          format: uuid
    RequestTitleUpdate:
      type: object
      properties:
        title:
          type: string
...